Set up a base qwerty layout
I really enjoy the arrow key setup of the Pok3r and have taken to having
a modifier under my left thumb without moving. This is a first attempt,
with a fair bit of duplication, to accommodate normal typing and
programming, making all of the symbols and numbers available below the
top row (thinking somewhat about Let's Split compatibility).

It is not quite symmetrical, but both raise and lower are easily reached
with either hand, and gui/alt are side by side on the left hand for easy
modifier chording. For usual typing, travel to the inside and outside
columns is pretty well minimized. I might try to go back to the default
of having shift on the home row for the left hand, but I have a ton of
muscle memory from putting control on caps lock for years; it was quite
difficult with the defaults when I tried it at first.