Set up Workman JK over my base qwerty layout
Workman JK is a cute name I came up for a small modification: swapping
only J and K for better Vim mechanics. ZXCV on the bottom row is not a
big muscle memory issue for me, so I'd consider swapping B and V as
well, but there was no pressing reason one way or the other, so I have
left them for now. With the arrows under my fingers at all times, I will
probably use them more, and might switch back to standard Workman.
We'll see how that goes.