Move backspace/enter/quotes around again
I've been having trouble with symbols on the right edge sharing keys
with enter or backspace, going back and forth between which I want on
the main layer.

This is a different approach, similar to one I tried briefly while space
and enter were still both on the right thumb. This time, the space
remains constant and both enter and backspace move onto the symbol
layer, on comma and period. This leaves the right edge for the quote and
minus symbols on main layer, shifted on the symbol layer.

I'm finding that I am making many fewer errors when looking to correct
or commit something, and effectively none with the quotes or dashes.
This also moves the greater and less than signs to the left hand, which
still keeps them handy for ERB. The equal and backslash placement seems
better: they are prioritized on the index fingers, making shifting easy
with either hand.